Allow typing only year in letter search date range ref #14
