+++ src/components/DateRangeFacet.jsx
@@ -1,6 +1,5 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { EuiTitle } from "@elastic/eui";
-import { useSearchkit } from "@searchkit/client";
 import moment from "moment";
 import { LetterDateFilter } from "./LetterDateFilter";
 import { datesValid } from "../common";
@@ -13,72 +12,13 @@ import { datesValid } from "../common";
  * @param {moment.Moment} props.maxDate Max date returned from API, as a moment object
  * @param {moment.Moment} props.minDate Min date returned from API, as a moment object
  * @param {boolean} props.loading Loading indicator boolean
+ * @param {object} props.dateRange Current date range state
+ * @param {Function} props.setDateRange Function to set date range state
  * @returns {React.Component} Date range React component
  */
-function DateRangeFacet({ minDate, maxDate, loading }) {
-    // adapted from @searchkit/elastic-ui
-    const api = useSearchkit();
-    const startDateFilters = api.getFiltersByIdentifier("start_date");
-    const selectedStartDate = startDateFilters && startDateFilters[0];
-    const endDateFilters = api.getFiltersByIdentifier("end_date");
-    const selectedEndDate = endDateFilters && endDateFilters[0];
-
-    // if a filter already present on mount, set initial state to that range
-    const [dateRange, setDateRange] = useState({
-        startDate: selectedStartDate?.dateMin
-            ? moment(selectedStartDate.dateMin)
-            : null,
-        endDate: selectedEndDate?.dateMax
-            ? moment(selectedEndDate.dateMax)
-            : null,
-    });
-
-    // update filters when range is changed
-    useEffect(() => {
-        api.removeFiltersByIdentifier("start_date");
-        api.removeFiltersByIdentifier("end_date");
-        if (
-            dateRange
-            && (dateRange.startDate || dateRange.endDate)
-            && datesValid({ ...dateRange })
-        ) {
-            if (dateRange.startDate) {
-                api.addFilter({
-                    identifier: "start_date",
-                    dateMin: dateRange?.startDate?.toISOString(),
-                });
-            }
-            if (dateRange.endDate) {
-                api.addFilter({
-                    identifier: "end_date",
-                    dateMax: dateRange?.endDate?.toISOString(),
-                });
-            }
-        }
-        api.search();
-    }, [dateRange]);
-
-    useEffect(() => {
-        // handle search state generated from URL query params
-        if (
-            (selectedStartDate?.dateMin && !dateRange.startDate)
-            || (selectedEndDate?.dateMax && !dateRange.endDate)
-        ) {
-            setDateRange(() => ({
-                startDate: selectedStartDate?.dateMin
-                    ? moment(selectedStartDate.dateMin)
-                    : null,
-                endDate: selectedEndDate?.dateMax
-                    ? moment(selectedEndDate.dateMax)
-                    : null,
-            }));
-            api.search();
-        } else if (!api.getFilters().length && (dateRange.startDate || dateRange.endDate)) {
-            // handle "reset search" click
-            setDateRange({ startDate: null, endDate: null });
-        }
-    }, [selectedStartDate, selectedEndDate]);
-
+function DateRangeFacet({
+    minDate, maxDate, dateRange, setDateRange, loading,
+}) {
     return (
         <>
             <EuiTitle size="xxs">
@@ -86,7 +26,9 @@ function DateRangeFacet({ minDate, maxDate, loading }) {
             </EuiTitle>
             <LetterDateFilter
                 dateRange={dateRange}
-                isValid={datesValid({ ...dateRange })}
+                isValid={datesValid({
+                    ...dateRange,
+                })}
                 loading={loading}
                 minDate={minDate}
                 maxDate={maxDate}
