+++ src/components/LetterDateFilter.jsx
@@ -30,6 +30,7 @@ export function LetterDateFilter({
                 <EuiDatePicker
                     aria-label="Start date"
                     endDate={dateRange?.endDate}
+                    dateFormat={["MM/DD/YYYY", "YYYY"]}
                     isInvalid={!isValid}
                     isLoading={loading}
                     minDate={minDate}
@@ -46,6 +47,7 @@ export function LetterDateFilter({
                 <EuiDatePicker
                     aria-label="End date"
                     endDate={dateRange?.endDate}
+                    dateFormat={["MM/DD/YYYY", "YYYY"]}
                     isInvalid={!isValid}
                     isLoading={loading}
                     minDate={minDate}
