+++ src/pages/LettersSearchPage/LettersSearchPage.jsx
@@ -40,6 +40,7 @@ import withSearchRouting from "../../components/withSearchRouting";
 import "../../common/search.css";
 import { SearchControls } from "../../components/SearchControls";
 import { useCustomSearchkitSDK } from "../../common";
+import { useDateFilter } from "./useDateFilter";
 
 // icon component cache required for dynamically imported EUI icons in Vite;
 // see https://github.com/elastic/eui/issues/5463
@@ -62,6 +63,7 @@ appendIconComponentCache({
 function LettersSearch() {
     const [query, setQuery] = useSearchkitQueryValue();
     const [operator, setOperator] = useState("or");
+    const [dateRangeState, setDateRangeState] = useDateFilter();
     const [sortState, setSortState] = useState({
         field: "date",
         direction: 1,
@@ -122,6 +124,8 @@ function LettersSearch() {
                             minDate={dateRange?.minDate}
                             maxDate={dateRange?.maxDate}
                             loading={dateRangeLoading || loading}
+                            dateRange={dateRangeState}
+                            setDateRange={setDateRangeState}
                         />
                         <EuiSpacer size="l" />
                         {results?.facets
@@ -164,7 +168,15 @@ function LettersSearch() {
                             </EuiTitle>
                         </EuiPageHeaderSection>
                         <EuiPageHeaderSection>
-                            <ResetSearchButton loading={loading} />
+                            <ResetSearchButton
+                                loading={loading}
+                                onClick={() => {
+                                    // customize reset behavior to also reset date range
+                                    api.setQuery("");
+                                    setDateRangeState({ startDate: null, endDate: null });
+                                    api.search();
+                                }}
+                            />
                         </EuiPageHeaderSection>
                     </EuiPageHeader>
                     <EuiPageContent>
